# internal/delivery/http/docs/swagger.yaml
definitions:
  models.CookingSteps:
    properties:
      cooking_time:
        type: number
      description:
        type: string
      guid:
        type: string
      order_number:
        type: integer
      recipe_id:
        type: string
    type: object
  models.CookingStepsUpdate:
    properties:
      cooking_time:
        type: number
      description:
        type: string
      guid:
        type: string
      order_number:
        type: integer
      recipe_id:
        type: string
    type: object
  models.CreateAggregatorRequest:
    properties:
      cooking_steps:
        items:
          $ref: '#/definitions/models.CreateCookingStepsRequest'
        type: array
      ingredients:
        items:
          $ref: '#/definitions/models.CreateIngredientsRequest'
        type: array
      recipe:
        $ref: '#/definitions/models.CreateRecipeRequest'
    type: object
  models.CreateAggregatorResponse:
    properties:
      guid:
        type: string
    type: object
  models.CreateCookingStepsRequest:
    properties:
      cooking_time:
        type: number
      description:
        type: string
      order_number:
        type: integer
    type: object
...
  models.CreateIngredientsRequest:
    properties:
      count:
        type: number
      dimension:
        type: string
      name:
        type: string
    type: object
  models.CreateRecipeRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
  models.CulinaryAggregator:
    properties:
      cooking_steps:
        items:
          $ref: '#/definitions/models.CookingSteps'
        type: array
      ingredients:
        items:
          $ref: '#/definitions/models.Ingredients'
        type: array
      recipe:
        $ref: '#/definitions/models.Recipe'
    type: object
...
  models.DeleteRecipeResponse:
    properties:
      succress:
        type: boolean
    type: object
...
  models.ErrorBadRequest:
    properties:
      error_code:
        type: integer
      message:
        type: string
    type: object
  models.Ingredients:
    properties:
      count:
        type: number
      dimension:
        type: string
      guid:
        type: string
      name:
        type: string
    type: object
  models.IngredientsUpdate:
    properties:
      count:
        type: number
      dimension:
        type: string
      guid:
        type: string
      name:
        type: string
    type: object
...
  models.Recipe:
    properties:
      cooking_time:
        type: number
      created_at:
        type: string
      description:
        type: string
      guid:
        type: string
      number_of_ratings:
        type: integer
      overall_rating:
        type: number
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.RecipeRatingRequest:
    properties:
      rating:
        type: integer
      recipe_id:
        type: string
    type: object
  models.RecipeRatingResponse:
    properties:
      succress:
        type: boolean
    type: object
  models.RecipeUpdate:
    properties:
      description:
        type: string
      guid:
        type: string
      title:
        type: string
    type: object
...
  models.UpdateRecipeRequest:
    properties:
      cooking_steps:
        items:
          $ref: '#/definitions/models.CookingStepsUpdate'
        type: array
      ingredients:
        items:
          $ref: '#/definitions/models.IngredientsUpdate'
        type: array
      recipe:
        $ref: '#/definitions/models.RecipeUpdate'
    type: object
  models.UpdateRecipeResponse:
    properties:
      succress:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /v1/recipe:
    get:
      consumes:
      - application/json
      description: Recipe List
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: cooking_time_from
        in: query
        name: cooking_time_from
        type: string
      - description: cooking_time_to
        in: query
        name: cooking_time_to
        type: string
      - description: guid,guid
        in: query
        name: ingridients
        type: string
      - description: rating_from
        in: query
        name: rating_from
        type: number
      - description: rating_to
        in: query
        name: rating_to
        type: number
      - description: order_rating
        in: query
        name: order_rating
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      summary: Recipe List
      tags:
      - aggregator
    post:
      consumes:
      - application/json
      description: Create recipe
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateAggregatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateAggregatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Create recipe
      tags:
      - aggregator
    put:
      consumes:
      - application/json
      description: Update recipe
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRecipeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateRecipeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Update recipe
      tags:
      - aggregator
...
  /v1/recipe/{id}:
    delete:
      consumes:
      - application/json
      description: Delete recipe by id
      parameters:
      - description: recipe_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeleteRecipeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Delete recipe by id
      tags:
      - aggregator
    get:
      consumes:
      - application/json
      description: Recipe by id
      parameters:
      - description: recipe_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CulinaryAggregator'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      summary: Recipe by id
      tags:
      - aggregator
  /v1/recipe/rating:
    post:
      consumes:
      - application/json
      description: Recipe rating
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RecipeRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecipeRatingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Recipe rating
      tags:
      - rating
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"